refactor(index.ts): expand exports for utils, types, and test mocks
Broadened the exports in src/index.ts to include additional utility modules,
type definitions, and test mock files. Also updated type exports to reference
individual files and directories. Renamed types/workers.ts to types/workers/index.ts
for consistency.
