+++ src/types/index.ts
@@ -8,3 +8,9 @@ export * from './modals';
 export * from './user-accounts';
 export * from './soap';
 export * from './folder';
+export * from './tags';
+export * from './sidebar';
+export * from './actions';
+export * from './identities';
+export * from './select';
+export * from './workers';
